fix: inconsistent txFirstSpend
Fixes a rare situation where, due to the parallel processing of transactions, it can happen that an input script is serialized as a wrapped script in a later transaction before it is serialized as a normal input script in a previous transaction. This is a simpler fix than moving the input processing of wrapped scripts to the last stage of the processing pipeline.

Reparsing the chain is recommended.